First changes from code review
